requirements.txt
apache-airflow-providers-amazon
boto3
kaggle
kagglehub
python-dotenv
